chore(docker-config): update docker support and ledger number
