+++ README.md
@@ -37,6 +37,7 @@ with the following system dependencies:
 To run via Docker, set up your docker environment as necessary and:
 
 ```bash
+$ make build
 $ make
 ```
