+++ init_accounts.py
@@ -1,38 +1,21 @@
 import logging
-import os
 import os.path
 
-import bigchaindb
 import bigchaindb.config_utils
 
 import apps_config
+from server.config_bigchaindb import get_bigchain
 from server.lib.models.accounts import Account
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-try:
-    CONFIG_FILE = os.environ['BIGCHAINDB_CONFIG']
-except KeyError:
-    CONFIG_FILE = os.path.join(os.path.dirname(__file__), '.bigchaindb_examples')
-
 APPS = apps_config.APPS
-
 
 LEDGER_API_BASE_HOST = os.environ.get('DOCKER_MACHINE_IP') or 'localhost'
 LEDGER_API_BASE_PORT = int(os.environ.get('LEDGER_API_BASE_PORT', '8000'))
 LEDGER_WS_BASE_HOST = os.environ.get('DOCKER_MACHINE_IP') or 'localhost'
 LEDGER_WS_BASE_PORT = int(os.environ.get('LEDGER_WS_BASE_PORT', '8888'))
-
-
-def get_bigchain(conf=CONFIG_FILE, ledger_id=None):
-    if os.path.isfile(conf):
-        bigchaindb.config_utils.autoconfigure(filename=conf, force=True)
-
-    if ledger_id is not None:
-        return bigchaindb.Bigchain(dbname='bigchaindb_examples_{}'.format(ledger_id))
-    else:
-        return bigchaindb.Bigchain()
 
 bigchain = get_bigchain()
 logging.info('INIT: bigchain initialized with database: {}'.format(bigchaindb.config['database']['name']))
